loa api

+++ internal/controller/variant.go
@@ -412,80 +412,94 @@ func GetVariantDetailsByDate(w http.ResponseWriter, r *http.Request) {
 }
 
 func CreateVariant(w http.ResponseWriter, r *http.Request) {
+	apiName := "variant_create"
+	valid := false
+
 	res := NewResponse(nil)
 	status := http.StatusUnauthorized
-	err := model.ErrTokenNotFound
-	errTitle := model.ErrCodeInvalidToken
+	err := model.ErrInvalidRole
+	errTitle := model.ErrCodeInvalidRole
 	res.AddError(its(status), errTitle, err.Error(), "Create Variant")
 
 	a := AuthToken(w, r)
 	if a.Valid {
-		status = http.StatusCreated
-
-		var rd Variant
-		decoder := json.NewDecoder(r.Body)
-		if err := decoder.Decode(&rd); err != nil {
-			log.Panic(err)
+		for _, valueRole := range a.User.Role {
+			features := model.ApiFeatures[valueRole.RoleDetail]
+			for _, valueFeature := range features {
+				if apiName == valueFeature {
+					valid = true
+				}
+			}
 		}
 
-		ts, err := time.Parse("01/02/2006", rd.StartDate)
-		if err != nil {
-			log.Panic(err)
-		}
-		te, err := time.Parse("01/02/2006", rd.EndDate)
-		if err != nil {
-			log.Panic(err)
-		}
-		fmt.Println(rd.ValidVoucherStart)
-		tvs, err := time.Parse("01/02/2006", rd.ValidVoucherStart)
-		if err != nil {
-			log.Panic(err)
-		}
-		tve, err := time.Parse("01/02/2006", rd.ValidVoucherEnd)
-		if err != nil {
-			log.Panic(err)
-		}
+		if valid {
+			status = http.StatusCreated
 
-		vr := model.VariantReq{
-			AccountId:          a.User.AccountID,
-			VariantName:        rd.VariantName,
-			VariantType:        rd.VariantType,
-			VoucherType:        rd.VoucherType,
-			VoucherPrice:       rd.VoucherPrice,
-			MaxQuantityVoucher: rd.MaxQuantityVoucher,
-			MaxUsageVoucher:    rd.MaxUsageVoucher,
-			AllowAccumulative:  rd.AllowAccumulative,
-			RedeemtionMethod:   rd.RedeemtionMethod,
-			DiscountValue:      rd.DiscountValue,
-			StartDate:          ts.Format("2006-01-02 15:04:05.000"),
-			EndDate:            te.Format("2006-01-02 15:04:05.000"),
-			StartHour:          rd.StartHour,
-			EndHour:            rd.EndHour,
-			ValidVoucherStart:  tvs.Format("2006-01-02 15:04:05.000"),
-			ValidVoucherEnd:    tve.Format("2006-01-02 15:04:05.000"),
-			VoucherLifetime:    rd.VoucherLifetime,
-			ValidityDays:       rd.ValidityDays,
-			ImgUrl:             rd.ImgUrl,
-			VariantTnc:         rd.VariantTnc,
-			VariantDescription: rd.VariantDescription,
-			ValidPartners:      rd.ValidPartners,
-		}
+			var rd Variant
+			decoder := json.NewDecoder(r.Body)
+			if err := decoder.Decode(&rd); err != nil {
+				log.Panic(err)
+			}
 
-		accountDetail, err := model.GetAccountDetailByUser(a.User.ID)
-		fr := model.FormatReq{
-			Prefix:     rd.VoucherFormat.Prefix,
-			Postfix:    accountDetail[0].Alias,
-			Body:       rd.VoucherFormat.Body,
-			FormatType: rd.VoucherFormat.FormatType,
-			Length:     rd.VoucherFormat.Length,
-		}
-		if id, err := model.InsertVariant(vr, fr, a.User.ID); err != nil {
-			//log.Panic(err)
-			status = http.StatusInternalServerError
-			errTitle = model.ErrCodeInternalError
-			res.AddError(its(status), errTitle, err.Error(), "Create Variant")
-		} else {
-			res = NewResponse(id)
+			ts, err := time.Parse("01/02/2006", rd.StartDate)
+			if err != nil {
+				log.Panic(err)
+			}
+			te, err := time.Parse("01/02/2006", rd.EndDate)
+			if err != nil {
+				log.Panic(err)
+			}
+			fmt.Println(rd.ValidVoucherStart)
+			tvs, err := time.Parse("01/02/2006", rd.ValidVoucherStart)
+			if err != nil {
+				log.Panic(err)
+			}
+			tve, err := time.Parse("01/02/2006", rd.ValidVoucherEnd)
+			if err != nil {
+				log.Panic(err)
+			}
+
+			vr := model.VariantReq{
+				AccountId:          a.User.AccountID,
+				VariantName:        rd.VariantName,
+				VariantType:        rd.VariantType,
+				VoucherType:        rd.VoucherType,
+				VoucherPrice:       rd.VoucherPrice,
+				MaxQuantityVoucher: rd.MaxQuantityVoucher,
+				MaxUsageVoucher:    rd.MaxUsageVoucher,
+				AllowAccumulative:  rd.AllowAccumulative,
+				RedeemtionMethod:   rd.RedeemtionMethod,
+				DiscountValue:      rd.DiscountValue,
+				StartDate:          ts.Format("2006-01-02 15:04:05.000"),
+				EndDate:            te.Format("2006-01-02 15:04:05.000"),
+				StartHour:          rd.StartHour,
+				EndHour:            rd.EndHour,
+				ValidVoucherStart:  tvs.Format("2006-01-02 15:04:05.000"),
+				ValidVoucherEnd:    tve.Format("2006-01-02 15:04:05.000"),
+				VoucherLifetime:    rd.VoucherLifetime,
+				ValidityDays:       rd.ValidityDays,
+				ImgUrl:             rd.ImgUrl,
+				VariantTnc:         rd.VariantTnc,
+				VariantDescription: rd.VariantDescription,
+				ValidPartners:      rd.ValidPartners,
+			}
+
+			accountDetail, err := model.GetAccountDetailByUser(a.User.ID)
+			fr := model.FormatReq{
+				Prefix:     rd.VoucherFormat.Prefix,
+				Postfix:    accountDetail[0].Alias,
+				Body:       rd.VoucherFormat.Body,
+				FormatType: rd.VoucherFormat.FormatType,
+				Length:     rd.VoucherFormat.Length,
+			}
+			if id, err := model.InsertVariant(vr, fr, a.User.ID); err != nil {
+				//log.Panic(err)
+				status = http.StatusInternalServerError
+				errTitle = model.ErrCodeInternalError
+				res.AddError(its(status), errTitle, err.Error(), "Create Variant")
+			} else {
+				res = NewResponse(id)
+			}
 		}
 	} else {
 		res = a.res
@@ -497,23 +511,42 @@ func CreateVariant(w http.ResponseWriter, r *http.Request) {
 
 func UpdateVariantRoute(w http.ResponseWriter, r *http.Request) {
 	types := r.FormValue("type")
+	apiName := "variant_update"
+	valid := false
 
 	res := NewResponse(nil)
 	status := http.StatusUnauthorized
-	err := model.ErrServerInternal
-	errTitle := model.ErrCodeInternalError
+	err := model.ErrInvalidRole
+	errTitle := model.ErrCodeInvalidRole
+	a := AuthToken(w, r)
+	if a.Valid {
+		for _, valueRole := range a.User.Role {
+			features := model.ApiFeatures[valueRole.RoleDetail]
+			for _, valueFeature := range features {
+				if apiName == valueFeature {
+					valid = true
+				}
+			}
+		}
+	}
+
 	if types == "" {
 		res.AddError(its(status), errTitle, err.Error(), "Update type not found")
 		render.JSON(w, res, status)
 	} else {
-		if types == "detail" {
-			UpdateVariant(w, r)
-		} else if types == "tenant" {
-			UpdateVariantTenant(w, r)
-		} else if types == "broadcast" {
-			UpdateVariantBroadcast(w, r)
+		if valid {
+			if types == "detail" {
+				UpdateVariant(w, r)
+			} else if types == "tenant" {
+				UpdateVariantTenant(w, r)
+			} else if types == "broadcast" {
+				UpdateVariantBroadcast(w, r)
+			} else {
+				res.AddError(its(status), errTitle, err.Error(), "Update type not allowed")
+				render.JSON(w, res, status)
+			}
 		} else {
-			res.AddError(its(status), errTitle, err.Error(), "Update type not allowed")
+			res.AddError(its(status), errTitle, err.Error(), "Update Variant")
 			render.JSON(w, res, status)
 		}
 	}
@@ -670,28 +703,44 @@ func UpdateVariantTenant(w http.ResponseWriter, r *http.Request) {
 }
 
 func DeleteVariant(w http.ResponseWriter, r *http.Request) {
-	fmt.Println("Delete Variant")
+	apiName := "variant_delete"
+	valid := false
+
 	res := NewResponse(nil)
 	id := r.FormValue("id")
 
 	status := http.StatusUnauthorized
-	errTitle := model.ErrCodeInvalidToken
+	err := model.ErrInvalidRole
+	errTitle := model.ErrCodeInvalidRole
+	res.AddError(its(status), errTitle, err.Error(), "Delete Variant")
 	a := AuthToken(w, r)
 	if a.Valid {
-		status = http.StatusOK
-		d := &model.DeleteVariantRequest{
-			Id:   id,
-			User: a.User.ID,
+		for _, valueRole := range a.User.Role {
+			features := model.ApiFeatures[valueRole.RoleDetail]
+			for _, valueFeature := range features {
+				if apiName == valueFeature {
+					valid = true
+				}
+			}
 		}
-		if err := d.Delete(); err != nil {
-			status = http.StatusInternalServerError
-			errTitle = model.ErrCodeInternalError
-			res.AddError(its(status), errTitle, err.Error(), "Delete Variant")
 
-		}
-		objName := strings.Split(d.Img_url, "/")
-		if deleteFile(w, r, objName[4]) {
-			return
+		if valid {
+			status = http.StatusOK
+			d := &model.DeleteVariantRequest{
+				Id:   id,
+				User: a.User.ID,
+			}
+			if err := d.Delete(); err != nil {
+				status = http.StatusInternalServerError
+				errTitle = model.ErrCodeInternalError
+				res.AddError(its(status), errTitle, err.Error(), "Delete Variant")
+
+			}
+			objName := strings.Split(d.Img_url, "/")
+			if deleteFile(w, r, objName[4]) {
+				return
+			}
+			res = NewResponse("")
 		}
 	} else {
 		res = a.res
